the easy stuff

+++ pets-lib/src/dialogue/dbox/mod.rs
@@ -3,6 +3,7 @@
 //!
 
 use dchoice::DChoice;
+use godot::builtin::Side;
 use godot::classes::{
     AnimationPlayer, Control, HBoxContainer, IPanelContainer, InputEvent,
     PanelContainer, RichTextLabel, Timer,
@@ -170,7 +171,7 @@ impl DialogBox {
     #[func]
     pub fn on_choice_picked(&mut self, choice: Gd<Control>) {
         // NOTE: convention is that the agent is BEFORE the labels
-        let picked_i = (choice.get_index() - 1) as u64;
+        let picked_i = choice.get_index() - 1;
         self.tween_choices_wave(false);
 
         self.base_mut()
